refactor(ivf_rabitq): merge ivf_rabitq into cuvs_objs
IVF-RaBitQ was the only neighbor algorithm built as a separate static
library and linked into libcuvs.so privately. This was inconsistent with
the rest of cuVS (IVF-PQ, CAGRA, IVF-Flat, etc., all compile directly
into the shared cuvs_objs target) and had two latent issues:

  - libcuvs_static.a did not include any IVF-RaBitQ implementation
    symbols, so static-only builds of cuVS would have failed to link
    against the IVF-RaBitQ public API.

  - The standalone target did not inherit CUVS_CUDA_FLAGS, which meant
    rapids_cuda_enable_fatbin_compression() never applied to
    IVF-RaBitQ AOT kernels. The IVF-RaBitQ portion of libcuvs.so's
    .nv_fatbin was effectively uncompressed.

Move the seven implementation files into cuvs_objs and drop the
ivf_rabitq target along with its bespoke target_compile_options. The
flags that were set per-target are all already covered by the
cuvs_objs configuration (--expt-extended-lambda, --expt-relaxed-constexpr,
OpenMP via OpenMP::OpenMP_CXX, -G for Debug via CUVS_DEBUG_CUDA_FLAGS).

The HIGH_ACC_FAST_SCAN compile definition was set unconditionally on
the standalone target and gated a single #if in quantizer_gpu.cuh
selecting NUM_SHORT_FACTORS = 1 vs 4. Since the macro was always
defined, the #else branch was dead. Inline NUM_SHORT_FACTORS = 1 and
drop the define so cuvs_objs does not pick up an IVF-RaBitQ-specific
flag globally.

Fix two -Werror=sign-compare diagnostics in ivf_gpu.cu that surfaced
once cuvs_objs's stricter warning flags applied to these files.

Drop "ivf_rabitq" from the link lines of two ANN benchmark targets in
cpp/bench/ann/CMakeLists.txt; libcuvs.so now exports those symbols
directly.

Binary size impact (multi-arch Release build):

  libcuvs.so          270 MB → 257 MB   (-13 MB, -4.8%)
  libcuvs.so .nv_fatbin: 217 MB → 193 MB (-24 MB, -11%)
  libcuvs_static.a    459 MB → 468 MB   (+9 MB; fix for missing
                                          IVF-RaBitQ symbols)
  libivf_rabitq.a     removed

+++ cpp/src/neighbors/ivf_rabitq/gpu_index/ivf_gpu.cu
@@ -366,7 +366,7 @@ void IVFGPU::save(const char* filename) const
   raft::copy(h_cluster_meta.data(), cluster_meta_.data_handle(), num_centroids, stream_);
   std::vector<size_t> cluster_sizes(num_centroids);
   raft::resource::sync_stream(handle_);
-  for (int i = 0; i < num_centroids; i++) {
+  for (size_t i = 0; i < num_centroids; i++) {
     cluster_sizes[i] = h_cluster_meta[i].num;
   }
   write_exact(cluster_sizes.data(), sizeof(size_t) * num_centroids);
@@ -807,7 +807,7 @@ void IVFGPU::construct_on_gpu_streaming(const float* host_data,
     // Use OpenMP parallel region to leverage persistent thread pool across batches
 #pragma omp parallel num_threads(num_threads)
     {
-      int tid           = omp_get_thread_num();
+      size_t tid        = omp_get_thread_num();
       size_t chunk_size = batch_vectors / num_threads;
       size_t start      = tid * chunk_size;
       size_t end        = (tid == num_threads - 1) ? batch_vectors : start + chunk_size;
